tello refactoring (from drivers to platforms)
