--- osgar/drivers/test_tello.py
@@ -1,13 +1,0 @@
-import unittest
-import datetime
-from unittest.mock import MagicMock
-
-from osgar.drivers.tello import TelloDrone
-
-
-class TelloDroneTest(unittest.TestCase):
-
-    def test_usage(self):
-        r = TelloDrone({}, MagicMock())
-
-# vim: expandtab sw=4 ts=4
